Refactor helpers a bit

+++ lib/reup.rb
@@ -8,9 +8,9 @@ require_relative "reup/helpers"
 include Reup::Helpers
 
 options = Slop.parse do |option|
-  option.bool "-r", "--resetdb", "resets the database"
-  option.string "-b", "--branch", "checkout branch"
-  option.on "-v", "--version", "print the version" do
+  option.bool   "-r", "--resetdb", "resets the database"
+  option.string "-b", "--branch",  "checkout branch"
+  option.on     "-v", "--version", "print the version" do
     puts Reup::VERSION
     exit
   end
@@ -19,12 +19,13 @@ end
 run "git checkout #{options[:branch]}", quiet: true if options.branch?
 run "git pull"
 
-if rails?
+case env
+when :rails
   run "bundle"
   run "rake db:migrate:reset" if options.resetdb?
-  serve
-elsif ember?
+when :ember
   run "npm install"
-  serve
 end
+
+serve
